fix (controls): add an offset after mouse movement
When moving the globe using GlobeControls, there was some kind of
offset, the globe not staying right under the cursor. This is visible if
the container of the canvas is not occupying the full page, the
container has a CSS position property different of the default 'static'
and there are elements right before it on the page, like the geoportal
case.

The solution to avoid this, is to detect the position of the container
and add an offset according to this position, using the
getBoundingClientRect method of the container.